requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2






/*Comandos*/

# Activar el entorno virtual
venv\Scripts\activate

# Instalar dependencias (solo la primera vez o en PC nuevo)
pip install -r requirements.txt

# Arrancar el servidor
python manage.py runserver


# Cuando cambias un modelo, crear la migración
python manage.py makemigrations

# Aplicar las migraciones a la BD
python manage.py migrate

# Abrir la consola de Django para hacer queries
python manage.py shell



# Importar jugadores desde el JS
python importar_jugadores.py

# Importar técnicas
python importar_tecnicas.py

# Importar equipos
python importar_equipos.py

# Crear superusuario para el admin
python manage.py createsuperuser

Admin:      http://127.0.0.1:8000/admin
Jugadores:  http://127.0.0.1:8000/api/personajes/
Técnicas:   http://127.0.0.1:8000/api/tecnicas/
Equipos:    http://127.0.0.1:8000/api/equipos/


from personajes.models import Personaje
Personaje.objects.filter(elemento='Monta\u00f1a').update(elemento='Montaña')

from tecnicas.models import Tecnica
Tecnica.objects.filter(elemento='Monta\u00f1a').update(elemento='Montaña')
